+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to
 ### Added
 
 - @cosmjs/math: Add `Decimal.floor` and `Decimal.ceil`.
+- @cosmjs/tendermint-rpc: Add `num_unconfirmed_txs` endpoint. ([#1150])
+
+[#1150]: https://github.com/cosmos/cosmjs/pull/1150
 
 ### Changed